refactor: harmonize text inputs structure in TextInputWrapper
